% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{dimer}
\alias{dimer}
\title{To construct a list with paramenters including     with a class}
\usage{
dimer(par = list(kon, koff, rmax))
}
\description{
To construct a list with paramenters including     with a class
}

