% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/equilibriumfit.R
\name{equilibrium_model}
\alias{equilibrium_model}
\title{Choose A Equilibrium Model.}
\usage{
equilibrium_model(model = c("simple1to1", "dimer"))
}
\arguments{
\item{model}{The model symbol. Choose the model bewteen "simple1to1", and "dimer"}
}
\description{
Choose a model between "simple1to1" and "dimer" to fit the equilibrium data
}

