% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfit.R
\name{kinfit}
\alias{kinfit}
\title{minimize the residFun using nls.lm}
\usage{
kinfit(init = NULL, dat = NULL, concs = concs, t2 = t2,
  model = "simple1to1", bound = NULL)
}
\arguments{
\item{dat}{the data in a list, which include a}

\item{model}{the model}

\item{par}{the initial parameters to be fitted}
}
\value{
the fitting results returned by minpack.lm::nls.lm
}
\description{
minimize the residFun using nls.lm
}
\examples{
# simulation
par = list(kon  = 2e2, koff = 1e-2, rmax = 1)
datsim = list()
datsim$concs <- 1e-5 * (2^(0:5));
datsim$xdata = seq(0, 300, length.out = 1501); # time
datsim$t2   = 150
datsim$ydata = NULL

}

