% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinsim.R
\name{kinsim}
\alias{kinsim}
\title{kinetic simulation}
\usage{
kinsim(par = par, model = "simple1to1", noise = 0.01)
}
\arguments{
\item{par}{parametes in a list, including kon, koff, rmax, cocns, t2, time}

\item{model}{the binding model, "simple1to1" or dimerization model : "dimer"}

\item{noise}{the noise level added to the simulated data}
}
\description{
kinetic simulation
}

