% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgplot.R
\name{plot.kinfit}
\alias{plot.kinfit}
\title{Plot A Sensorgram For A kinfit Object}
\usage{
\method{plot}{kinfit}(object, ...)
}
\description{
Plot the predicted responses (nm) vs the Time (sec) variable at the corresponding
concentrations of ligand in solution. The fitted parameters kon, koff, rmax are
also presented on the plot.
}

