% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinsim.R
\name{predict.kinfit}
\alias{predict.kinfit}
\title{Predict method for Kinetic Model Fits}
\usage{
\method{predict}{kinfit}(object, new_concs, noise = c(0, 0.01))
}
\arguments{
\item{object}{An kinfit object returned from \code{\link{kinfit}}.}

\item{newdata}{}
}
\description{
Predict the responses (nm) correponding to concentrations based on a kinfit object.
}

