% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{simple1to1}
\alias{simple1to1}
\title{provide a constructor function of S3 class "simple1to1"}
\usage{
simple1to1(kon, koff, rmax)
}
\description{
provide a constructor function of S3 class "simple1to1"
}

