% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kinsim.R
\name{kinsim}
\alias{kinsim}
\title{Kinetic Simulation of Protein-Ligand Sensorgram Data}
\usage{
kinsim(par, model = "simple1to1", noise = 0.01)
}
\arguments{
\item{par}{A list of parametes, including kon, koff, rmax, concs, t2, time}

\item{model}{The binding model, "simple1to1" or dimerization model : "dimer"}

\item{noise}{The noise level added to the simulated data}
}
\value{
kinsim returns an data frame. The first column of the data frame is
  "Time", which is the same the input variable time, and it is followed by the
  calculated responses corresponding to each concentration of the concs variable
}
\description{
kinsim generates a sensorgram data according to the input parameters, kon, koff,
rmax, concs, t2, time and the chosen model.
}
\examples{
# Simulation  ----------------------------------------------------------

# Prepare the parameters for simulation
par = list(kon    = 2e2,
          koff   = 1e-2,
          rmax   = 1,
          concs  = 1e-5 * (2^(0:5)),
          time   = seq(0, 300, length.out = 1501),
          t2     = 150)

# Simulating
model = "simple1to1"
xySimulated <- kinsim(par = par, model = model, noise = 0.01)
}
\seealso{
\code{\link{kinfit}}
}

