% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kinsim.R
\name{kinsim}
\alias{kinsim}
\title{kinetic simulation}
\usage{
kinsim(par, model = "simple1to1", noise = 0.01)
}
\arguments{
\item{par}{parametes in a list, including kon, koff, rmax, cocns, t2, time}

\item{model}{the binding model, "simple1to1" or dimerization model : "dimer"}

\item{noise}{the noise level added to the simulated data}
}
\description{
kinetic simulation
}
\examples{
# Simulation  ----------------------------------------------------------

# Prepare the parameters for simulation
par = list(kon    = 2e2,
          koff   = 1e-2,
          rmax   = 1,
          concs  = 1e-5 * (2^(0:5)),
          time   = seq(0, 300, length.out = 1501),
          t2     = 150)

# Simulating
model = "simple1to1"
xySimulated <- kinsim(par = par, model = model, noise = 0.01)
}
\seealso{
\code{\link{kinfit}}
}

