% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/equilibriumfit.R
\name{ssgfit}
\alias{ssgfit}
\title{Equilibrium Data Fitting of Protein-Ligand Sensorgram}
\usage{
ssgfit(dat, concs, start = list(Rmax = 0.5, Kd = 1), width = 5,
  t2 = 149.5, model = c("simple1to1", "dimer"))
}
\description{
Equilibrium Data Fitting of Protein-Ligand Sensorgram
}

